Updating makefile
mingw32-make clean does not found "del" command to remove old objects. However "rm" does is in the folder C:\mingw32\usr\bin. It is needed to update the video #1 to add this action. It is not critical if it does not matter to us to be unable to remove *.o files, but it is a little bit annoying.